Mods won't have fileSize, only modVersions
Because of some changes to the backend PRs, (the PR was split out into multiple), mods won't have a fileSize. So fileSize has to be taken from a modVersion.
This is better anyways, that way instead of always being the latest modVersion, it takes the latest modVersion available for a version. This makes the fileSize much more accurate if the mods size changes a lot between versions